requirements.txt
openai
Flask
packaging
python-dotenv
requests